user config file renamed
